[WFGP-222] Generate YAML channel definition
When a feature-pack is built, create a new artifact (with "channel"
classifier" and "yaml" extension) that represents the channel definition
for the feature pack.
This feature is controlled by the generate-channel
parameter (optional, false  by default)

Dependencies to other feature-packs are defined in the channel's
"requires" field. They can be skipped if
add-feature-packs-as-required-channels is set to false (optional, true
by default).

JIRA: https://issues.redhat.com/browse/WFGP-222

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>