[WFGP-298] Add copyright information when the feature pack is built
The copyright is used by the document generator to add a copyright footer to every
generated page.
This attribute is optional (and the copyright footer is not displayed if it is not set).
The year of the copyright is computed when the documentation is generated.

This fixes https://issues.redhat.com/browse/WFGP-298

Signed-off-by: Jeff Mesnil <jmesnil@ibm.com>